Optimize the image and vendor dependencies
This change aims at 1) optimizing the final image size, 2) allow
reproducible builds via dependency vendoring.

The first thing it does is to switch to ubi_/ubi-micro base
image. This reduces the final image size by 25 MB, i.e. 34 %.

The second thing is to use the -ldflags="-s -w" option of
the Go compiler to strip to binary. This reduces the binary size
from 14 MB to 11 MB, i.e. 25 %. In total, the image size is reduce by
44 %, from 89.3 MB to 49.4 MB.

This change also switches to the ubi8/go-toolset base image for the
builder stage. This is the image we recommend to the community.

This change also stores the Go dependencies in the Git repositories,
in order to be able to reproduce builds even when the dependency has
disappeared from public repositories. It also allows to more easily
dive in the dependencies when finding regressions.

Signed-off-by: Fabien Dupont <fabiendupont@pm.me>